+++ core-ui/src/components/Predefined/Create/DNSProviders/DNSProviders.create.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { ResourceForm } from '../../../../shared/ResourceForm/ResourceForm';
+import { ResourceForm } from 'shared/ResourceForm/ResourceForm';
 import { createDNSProviderTemplate } from './templates';
 import { ProviderTypeDropdown } from './ProviderTypeDropdown';
 import { SecretRef } from 'shared/components/ResourceRef/SecretRef';
