+++ core-ui/src/components/Predefined/Create/Deployments/Deployments.create.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { usePost, useNotification } from 'react-shared';
-import { ResourceForm } from '../../../../shared/ResourceForm/ResourceForm';
 import { Button, Checkbox } from 'fundamental-react';
 import * as jp from 'jsonpath';
-import './Deployments.create.scss';
 
+import { ResourceForm } from 'shared/ResourceForm/ResourceForm';
+import * as Inputs from 'shared/ResourceForm/components/Inputs';
+
+import './Deployments.create.scss';
 import {
   createContainerTemplate,
   createDeploymentTemplate,
@@ -144,7 +146,7 @@ export function DeploymentsCreate({
         simple
         propertyPath="$.spec.template.spec.containers[0].image"
         label={t('deployments.create-modal.simple.docker-image')}
-        input={ResourceForm.Input}
+        input={Inputs.Text}
         placeholder={t(
           'deployments.create-modal.simple.docker-image-placeholder',
         )}
@@ -190,7 +192,7 @@ export function DeploymentsCreate({
           disabled={!createService}
           propertyPath="$.spec.ports[0].port"
           label={t('deployments.create-modal.advanced.port')}
-          input={ResourceForm.Number}
+          input={Inputs.Number}
           placeholder={t('deployments.create-modal.advanced.port-placeholder')}
         />
         <ResourceForm.FormField
@@ -199,7 +201,7 @@ export function DeploymentsCreate({
           disabled={!createService}
           propertyPath="$.spec.ports[0].targetPort"
           label={t('deployments.create-modal.advanced.target-port')}
-          input={ResourceForm.Number}
+          input={Inputs.Number}
           placeholder={t(
             'deployments.create-modal.advanced.target-port-placeholder',
           )}
