+++ core-ui/src/shared/ResourceForm/ResourceForm.js
@@ -156,8 +156,6 @@ ResourceForm.Single = SingleForm;
 ResourceForm.Wrapper = ResourceFormWrapper;
 
 ResourceForm.Label = FormComponents.Label;
-ResourceForm.Input = FormComponents.Input;
-ResourceForm.Number = FormComponents.Number;
 ResourceForm.CollapsibleSection = FormComponents.CollapsibleSection;
 ResourceForm.FormField = FormComponents.FormField;
 ResourceForm.TextArrayInput = FormComponents.TextArrayInput;
