+++ core-ui/src/shared/ResourceForm/components/FormComponents.js
@@ -77,29 +77,6 @@ export function Label({ required, tooltipContent, children }) {
   }
 }
 
-export function Input({ value, setValue, ...props }) {
-  return (
-    <FormInput
-      compact
-      value={value}
-      onChange={e => setValue(e.target.value)}
-      {...props}
-    />
-  );
-}
-
-export function Number({ value, setValue, ...props }) {
-  return (
-    <FormInput
-      compact
-      type="number"
-      value={value}
-      onChange={e => setValue(e.target.valueAsNumber)}
-      {...props}
-    />
-  );
-}
-
 export function FormField({
   simple,
   advanced,
