+++ core-ui/src/shared/ResourceForm/components/Inputs.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { FormInput } from 'fundamental-react';
+
+export function Text({ value, setValue, ...props }) {
+  return (
+    <FormInput
+      compact
+      value={value}
+      onChange={e => setValue(e.target.value)}
+      {...props}
+    />
+  );
+}
+
+export function Number({ value, setValue, ...props }) {
+  return (
+    <FormInput
+      compact
+      type="number"
+      value={value}
+      onChange={e => setValue(e.target.valueAsNumber)}
+      {...props}
+    />
+  );
+}
